application/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Database (SQLite - no extra drivers needed)
SQLAlchemy==2.0.23
alembic==1.13.0

# Authentication & Security
Werkzeug==3.0.1
bcrypt==4.1.2
python-dotenv==1.0.0

# AI/ML - Core (updated versions)
torch>=2.2.0
transformers>=4.36.0
sentence-transformers>=2.2.0
accelerate>=0.25.0

# AI/ML - Fine-tuning
peft>=0.7.0
datasets>=2.16.0
scikit-learn>=1.3.0
bitsandbytes>=0.41.0  # For 8-bit LLaMA quantization

# Vector Search
faiss-cpu>=1.8.0
numpy>=1.24.0

# NLP Utilities
nltk>=3.8
spacy>=3.7

# File Processing
PyPDF2>=3.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'

# HTTP & API
requests>=2.31.0
urllib3>=2.0.0
huggingface_hub>=0.20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0

# Production Server
gunicorn>=21.2.0

# Utilities
click>=8.1.0
python-dateutil>=2.8.0
pytz>=2023.3

# Data Processing
pandas>=2.1.0
